edit prediction

+++ src/main/java/com/example/stocks/dto/PreResPredictedDto.java
@@ -1,11 +1,12 @@
 package com.example.stocks.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 // 예측값 리스트 인터페이스
 public interface PreResPredictedDto {
     String getShortCode();
-    Long getPreviousPrice();
-    Double getPreviousPercentage();
+    BigDecimal getPreviousPrice();
+    BigDecimal getPreviousPercentage();
     Date getPreviousDate();
 }
